+++ Addams.Tests/TestsModel.cs
@@ -8,34 +8,29 @@ namespace Addams.Tests;
 public class TestsModel
 {
     [Test]
-    public void TestDurationValue()
+    [TestCase(155000, "00:02:35")]
+    [TestCase(5747000, "01:35:47")]
+    public void TestDurationValue(int duration, string expected)
     {
-        // Arrange
-        string expected = "00:02:35";
-        int duration = 155000;
-
+        // Act
         TimeSpan ts = TimeConverter.ConvertMsToTimeSpan(duration);
         string time = TimeConverter.FormatTimeSpan(ts);
 
-        Assert.That(expected == time);
-
-        // Second test
-        expected = "01:35:47";
-        duration = 5747000;
-
-        ts = TimeConverter.ConvertMsToTimeSpan(duration);
-        time = TimeConverter.FormatTimeSpan(ts);
+        // Assert
         Assert.That(expected == time);
     }
 
     [Test]
     public void TestTrackDuration()
     {
+        // Arrange
         const string expected = "00:04:12";
         const int duration = 252000;
 
+        // Act
         Track track = new() { Duration = duration };
 
+        // Assert
         Assert.That(expected == track.DurationFormatted);
     }
 }